+++ CHANGELOG.md
@@ -6,10 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.1.3] - 2026-07-09
-- Default version for wingify smartcode nextjs is v3.
 
 ### Changed
-
+- Default version for wingify smartcode nextjs is v3.
 
 ## [1.1.2] - 2026-06-17