+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.1.3] - 2026-07-09
 
 ### Changed
-- Default version for wingify smartcode nextjs is v3.
+- Updated the default version for SmartCode to 3.0.
 
 ## [1.1.2] - 2026-06-17